+++ packages/serialize/lib/index.ts
@@ -1,4 +1,4 @@
-export { default as serialize } from "./serialize.middleware";
+export { default as serialize, SerializeOption } from "./serialize.middleware";
 export { default as SerializeTokens } from "./serialize.tokens";
 export {
   default as SerializerModule,
@@ -6,3 +6,4 @@ export {
 } from "./serializer.module";
 
 export { default as SerializeContext } from "./serialize.context";
+export { default as registerDefaultSerializer } from "./register-default-serializer";
